Run the local dispatcher from its own module
LocalDispatcher, the slot that accounts for its capacity, and the helper
that stamps a provider's credentials onto an agent config all lived in
main.rs, which is otherwise the file that decides what the binary is and
then stops. None of this is startup: the cap, the slot and the spawn all
happen on the request path, for the length of an interview, long after
that decision is made.

A move, not a rewrite. Every line is the one it was, and what it buys is
tests: the slot's release-by-drop and the credential inheritance now run
against the library rather than needing a process, which is why the two
tests move with them rather than staying behind in main.
